story base on preference is working

+++ Assets/Scripts/MatchmakingState.cs
@@ -7,7 +7,7 @@ public class MatchmakingState : UnitySingletonPersistent<MatchmakingState>
     public float compatibility = 0;
     public string name1 = "";
     public string name2 = "";
-    public Dictionary<string, float> sharedPreferences;
+    public Dictionary<string, int> sharedPreferences;
     //string : preferene name
     //Float: -1 both dislike, 0 one like one dislike, 1 both like
 }
